backend/requirements.txt
sentence-transformers>=2.2.0
torch>=1.13.0

fastapi>=0.109.0
uvicorn>=0.27.0
httpx>=0.26.0
python-jose[cryptography]>=3.3.0
python-dotenv>=1.0.0
sqlalchemy>=2.0.25
pydantic>=2.6.0
